Remove ESLint rule about a11y emoji, because...
...The lint rule isn't sophisticated enough to detect that I'm
following it!

I wrote a custom <Emoji> component which wraps it in a <span>,
gives it a 'role' of "img", and gives it an aria label. Sadly,
because this happens within the component, and the emoji is written
outside of it (and passed as children), the lint rule thinks I'm
violating it.

So yeah. Not actually breaking this rule :) but disabling it.