Refactor routes
Woot! Finally got around to doing this big routes refactor.

The core idea is that routes are nested now. A top-level <Switch>
handles the static pages, and the main landing page, but delegates
to <AppRoutes /> to handle desktop and mobile routing for the main
logged-in application.

For this to work, I had to create a <LoggedInLayout /> component,
used within <AppRoutes>. This 'inversion of control' feels right;
the layout handles the header and styling of the page.

Also added a convenience component, <AuthenticatedRoute />, which
renders a fallback component if the user isn't logged in.