Simplify GoogleButton, rename MobileLoggedOutMenu
It was suggested to me that the "Login" button in the top-right of
the landing page should open a modal, rather than be a direct Google
link.

I repurposed the MobileLoggedOutMenu to work for desktop as well,
and made it use the SignupButtons component already used.

Once that was done, I realized I was never not using the `official`
prop for GoogleButton, so I killed the non-official variant. I was
also sick of downloading a totally different image for the `pressed`
state, and so I recreated it using a simple :active CSS pseudoselector.